issue: 987687 Request buffer for next packet after send_ring_buffer()
For non-inline packets (size > ~180 bytes) we request the packet buffer
during the fast_send() call. This could be optimized as we do for inline
packets: request the buffer for the next packet after calling
send_ring_buffer() while the packet is sent to the wire.

This commit improves latency for non-inline, non-fragmented, UDP packets

Signed-off-by: Liran Oz <lirano@mellanox.com>