issue: 910917 Return error on socket API if exception occurred
Calling socket API may initiate VMA global constructors.
If a constructor fails there is an exception. Following this commit
the failure will be propagated to the calling API.
Returned value will be -1 and errno will indicate the error type.

This commit complements issue "vma crashes after installed on RH7.3
inbox driver" where the crash was avoided by catching an exception but
the failure was unknown to the calling API

Signed-off-by: Rafi Wiener <rafiw@mellanox.com>